backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.1
python-docx==1.1.0
python-pptx==0.6.23
firebase-admin==6.3.0
Pillow>=10.1.0
gunicorn==21.2.0

#### 5. Create `backend/Procfile`
**Location:** `backend/Procfile`
web: uvicorn main:app --host 0.0.0.0 --port $PORT